--- codex-rs/core/src/exec_command/mod.rs
@@ -1,3 +1,0 @@
-mod exec_command_session;
-
-pub(crate) use exec_command_session::ExecCommandSession;
