+++ codex-rs/core/src/tools/runtimes/apply_patch.rs
@@ -8,17 +8,15 @@ use crate::CODEX_APPLY_PATCH_ARG1;
 use crate::exec::ExecToolCallOutput;
 use crate::sandboxing::CommandSpec;
 use crate::sandboxing::execute_env;
-use crate::tools::sandboxing::{
-    Approvable,
-    ApprovalCtx,
-    ApprovalDecision,
-    SandboxAttempt,
-    Sandboxable,
-    SandboxablePreference,
-    ToolCtx,
-    ToolError,
-    ToolRuntime,
-};
+use crate::tools::sandboxing::Approvable;
+use crate::tools::sandboxing::ApprovalCtx;
+use crate::tools::sandboxing::ApprovalDecision;
+use crate::tools::sandboxing::SandboxAttempt;
+use crate::tools::sandboxing::Sandboxable;
+use crate::tools::sandboxing::SandboxablePreference;
+use crate::tools::sandboxing::ToolCtx;
+use crate::tools::sandboxing::ToolError;
+use crate::tools::sandboxing::ToolRuntime;
 use std::collections::HashMap;
 use std::path::PathBuf;
