+++ codex-rs/core/src/tools/runtimes/command_spec.rs
@@ -1,5 +1,5 @@
-use crate::tools::sandboxing::ToolError;
 use crate::sandboxing::CommandSpec;
+use crate::tools::sandboxing::ToolError;
 use std::collections::HashMap;
 use std::path::PathBuf;
